powerpointfonts_web/requirements.txt
flask>=2.0.0
python-pptx>=1.0.0
werkzeug>=2.0.0 